Full API documentation

Gradus.tracegeodesics — Function

tracegeodesics(
    m::AbstractMetricParams{T},
    position,
    velocity,
    time_domain::Tuple{T,T};
    solver = Tsit5(),
    μ = 0.0,
    closest_approach = 1.01,
    effective_infinity = 1200.0,
    callback = nothing,
    solver_opts...,
)

Integrate a geodesic for metric parameterised by m, for some initial positions and velocities. The positions and velocities may be

a single position and velocity in the form of a vector of numbers,

a collection of positions and velocities, as either a vector of vectors, or as a matrix,

a single position and a function with signature $vel_func(i)$ which returns a four-velocity.

The matrix specification reads each corresponding column as the initial position and velocity. When a collection of positions and velocities is supplied, this method dispatched EnsembleProblem, offering ensemble as a solver_opts, specifying the ensemble method to use.

solver_opts are the common solver options in DifferentialEquations.

Gradus.GradusBase.inverse_metric_components — Function

inverse_metric_components(g_comp)

Calculates $g^{tt}$, $g^{rr}$, $g^{\theta\theta}$, $g^{\phi\phi}$, $g^{t\phi}$ of a static, axis-symmetric metric from $g_{tt}$, $g_{rr}$, $g_{\theta\theta}$, $g_{\phi\phi}$, $g_{t\phi}$ using a symbolically computed inverse matrix method.

Notes

To recreate:

using Symbolics
@variables g[1:5] # non zero metric components
metric = [
    g[1] 0 0 g[5]
    0 g[2] 0 0
    0 0 g[3] 0
    g[5] 0 0 g[4]
]
inv(metric)

Gradus.metric_jacobian — Function

metric_jacobian(m::AbstractAutoDiffStaticAxisSymmetricParams{T}, rθ)

Calculate the Jacobian elements of the metric with respect to $r$ and $\theta$.

Limitations:

currenly pre-supposes static, axis-symmetric metric.

Gradus.intersects_geometry — Function

intersects_geometry(m::AbstractAccretionGeometry{T}, line_element)

Utility function. Returns a boolean dependent on whether line_element intersects with the geometry (true) or not (false). Uses in_nearby_region to optimze and calls has_intersect to determine intersection.

Gradus.build_collision_callback — Function

build_collision_callback(m::AbstractAccretionGeometry{T})

Generates the callback used for the integration. Returns a Function, with the fingerprint

function callback(u, λ, integrator)::Bool
    # ...
end

Gradus.jsf_algorithm — Function

jsf_algorithm(V₁::T, V₂::T, V₃::T, Q₁::V, Q₂::V; ϵ = 1e-8)

Implemented from Jiménez, Segura, Feito. Computation Geometry 43 (2010) 474-492.

See this blog post for a discussion.