GradusBase API documentation

Gradus.GradusBase.E — Method

E(m::AbstractMatrix{T}, v)
E(m::AbstractMetric{T}, u, v)

Compute the energy for a numerically evaluated metric, and some velocity four vector v,

\[E = - p_t = - g_{t\nu} p^\nu.\]

For null geodesics, the velocity is the momentum $v^\nu = p^\nu$. For massive geodesics, the mass $\mu$ needs to be known to compute $\mu v^\nu = p^\nu$.

Gradus.GradusBase.Lz — Method

Lz(m::AbstractMatrix{T}, v)
Lz(m::AbstractMetric{T}, u, v)

Compute the angular momentum for a numerically evaluated metric, and some velocity four vector v.

\[L_z = p_\phi = - g_{\phi\nu} p^\nu.\]

Gradus.GradusBase._tetrad_permute — Method

_tetrad_permute(x)

Permute the spatial components (vectors) of a given tetrad. That is, given some x = (e1, e2, e3, e4), a single call to _tetrad_permute will return (e1, e4, e2, e3).

Attempts to respect the input type in the return type. Checked for SVector and NTuple.

Gradus.GradusBase.constrain — Method

constrain(m::AbstractMetric, x, v; μ=0)

Calculate the time component $v^t$ of a velocity vector v, which would constrain the vector at a position x as a geodesic with invariant mass μ.

The velocity vector needs to only specify the $v^r$, $v^\theta$, and $v^\phi$ component, as the $v^t$ is constrained in GR by

\[g_{\sigma\nu} v^\sigma v^\nu = -\mu^2,\]

where $\mu^2$ is the invariant mass of the particle. This furthermore permits a choice of geodesic to trace. The choices correspond to

μ = 0.0 (default): null geodesic

μ > 0.0: time-like geodesic

μ < 0.0: space-like geodesic

Gradus.GradusBase.geodesic_equation — Method

geodesic_equation(m::AbstractMetric, x, v)

Calculate the four-acceleration of the geodesic equation for a spacetime given by the metric m, four-position x and four-velocity v.

A geodesic is the shortest path connecting two points in space. For flat space, this is just a straight line. In curved space, geodesics are analogous to straight lines between points (e.g. the great circle on a sphere).

The geodesic equation calculates the acceleration experienced by a particle at position $x^\mu = (t, r, \theta, \phi)$ travelling with tangential velocity $v^\nu = \text{d} x / \text{d} \lambda$ due to the curvature of spacetime. The curvature is calculated from the metric, encoded in the Christoffel symbols. The acceleration is then calculated via

\[\frac{\text{d}^2 x^\mu}{\text{d} \lambda^2} = - \Gamma^{\mu}_{\phantom{\mu}\nu\sigma} \frac{\text{d}x^\nu}{\text{d} \lambda} \frac{\text{d}x^\sigma}{\text{d} \lambda}\]

where $\Gamma^{\mu}_{\phantom{\mu}\nu\sigma}$ are the Christoffel symbols (of the second kind), and $\lambda$ is an affine parameter that parameterizes the solution.

Gradus.GradusBase.inner_radius — Method

inner_radius(m::AbstractMetric{T})

Return the innermost valid coordinate relative to the origin, for use in geodesic tracing.

This usually represents some property of the metric, e.g. event horizon radius in Kerr/Schwarzschild metrics, or throat diameter in worm hole metrics.

Gradus.GradusBase.metric — Method

metric(m::AbstractMetric{T}, u)

Numerically evaluate the corresponding metric for AbstractMetric, given parameter values m and some point u.

Gradus.GradusBase.metric_components — Method

metric_components(m::AbstractMetric{T}, x)

Return a tuple with each non-zero metric component for the metric described by m at position x. Note that the position need not be a four-vector, and for specific implementations may only be a subset of the total manifold coordinates. See specific implementations for subtypes of AbstractMetric for details.

Gradus.GradusBase.metric_type — Method

metric_type(m::AbstractMetric{T})

Return the AbstractMetric type associated with the metric parameters m.

Gradus.GradusBase.mproject — Method

mproject(g, v, u)

Project vector v onto u with metric g. Optional first argument may be AbstractMetric for more optimized methods, which fallback to an einsum.

Gradus.GradusBase.tetradframe — Method

tetradframe(m::AbstractMetric, x, v)
tetradframe(g::AbstractMatrix [= metric(m, x)], v)

Compute a tetrad frame via the Gram-Schmidt orthonormalization procedure (see gramschmidt), where the first vector of the frame is the velocity vector v at some position x.

Used to compute the locally non-rotating frame via lnrframe and lnrbasis respectively.

Attempts to reorder the tetrad into a canonical form associated with the global coordinates, that is, returns a tuple that corresponds to the $x^1, x^2, x^3, x^4$ coordinates of the spacetime.

Gradus.GradusBase.unpack_solution — Method

unpack_solution([m], sol)

Unpack a solution (SciMLBase.AbstractODESolution) as a GeodesicPoint, optionally specifying the metric under which quantities are transformed.

If the solution stores any additional parameters (e.g. intensity in radiative transfer), these will be packed into the aux field of GeodesicPoint.

Example use

sol = tracegeodesics(m, x, v)
point = unpack_solution(sol)

Gradus.GradusBase.unpack_solution_full — Method

unpack_solution_full

Unpacks each point in the solution, similar to unpack_solution but returns an array of GeodesicPoint.

Gradus.GradusBase.AbstractGeodesicPoint — Type

abstract type AbstractGeodesicPoint

Supertype for geodesic points, used to store information about specific points along geodesic trajectories.

Note

Currently limited to storing the start and endpoint of any given trajectory. To keep the full geodesic path, it is encouraged to use the SciMLBase.AbstractODESolution directly.

Must minimally have the same fields as GeodesicPoint. Examples include Gradus.FirstOrderGeodesicPoint.

Gradus.GradusBase.AbstractIntegrationParameters — Type

AbstractIntegrationParameters

Parameters that are made available at each step of the integration, that need not be constant. For example, the turning points or withing-geometry flags.

Gradus.GradusBase.AbstractMetric — Type

abstract type AbstractMetric{T} end

Abstract type used to dispatch different geodesic problems.

Gradus.GradusBase.AbstractTrace — Type

AbstractTrace

Parameters that are constant throughout the integration (e.g. mass or frequency) for any number of geodesics. Also used to dispatch different tracing problems.

Gradus.GradusBase.GeodesicPoint — Type

struct GeodesicPoint <: AbstractGeodesicPoint

status: Return code of the integrator for this geodesic.

λ_min: Start affine time

λ_max: End affine time

x_init: Start four position

x: End four position

v_init: Start four velocity

v: End four velocity

aux: Auxillary values (polarisation, intensities, etc.)