Defining a new metric

Gradus.jl is able to integrate any 3+1 dimensional metric, however provides a few derivative abstract types to implement to ensure the most efficient code is executed for a given metric. Currently, Gradus differentiates the following specializations:

First-Order

Gradus.FirstOrderMethods.AbstractFirstOrderMetricParams — Type

AbstractFirstOrderMetricParams{T} <: AbstractMetricParams{T}

Abstract type for metrics using the 1st-order integration method. The 1st-order methods reuse the velocity vector as a parameter vector, where only element vel[2] and vel[3] are used, and are local observer ratios $\sin \Theta$ and $\sin \Phi$ respectively.

Require implementation of

GradusBase.inner_radius

GradusBase.constrain

FirstOrderMethods.four_velocity

FirstOrderMethods.calc_lq

FirstOrderMethods.Vr

FirstOrderMethods.Vθ

GeodesicTracer.alpha_beta_to_vel

Gradus.FirstOrderMethods.four_velocity — Function

four_velocity(u, m::AbstractFirstOrderMetricParams{T}, p) -> NTuple{4, Any}

Calculate the four-velocity at a point u, given a set of metric parameters and the constants of motion in p.

Gradus.FirstOrderMethods.calc_lq — Function

Calculate constants of motion $L$ and $Q$, given a set of metric parameters, the geodesic position, and the param vector.

Gradus.FirstOrderMethods.Vr — Function

Vr(m::AbstractFirstOrderMetricParams{T}, u, p) -> Any

Effective potential in the radial direction. Used only to track sign changes.

Gradus.FirstOrderMethods.Vθ — Function

Effective potential in the angular direction. Used only to track sign changes.

Second-Order

Gradus.GeodesicTracer.AbstractAutoDiffMetricParams — Type

AbstractAutoDiffMetricParams{T} <: AbstractMetricParams{T}

Abstract type for metrics using the 2nd-order integration method, with the automatic differentiation backend.

Gradus.GeodesicTracer.metric_components — Function

metric_components(m::AbstractAutoDiffStaticAxisSymmetricParams{T}, rθ) -> StaticArrays.SVector{5}

Interface for GeodesicTracer.AbstractAutoDiffStaticAxisSymmetricParams. Should return a vector or tuple with the elements

\[\left( g_{tt}, g_{rr}, g_{\theta \theta}, g_{\phi \phi}, g_{t\phi} \right).\]

Gradus.GeodesicTracer.AbstractAutoDiffStaticAxisSymmetricParams — Type

AbstractAutoDiffStaticAxisSymmetricParams{T} <: AbstractAutoDiffMetricParams{T}

Specialisation for static, axis-symmetric metrics. Here, the metric is of the form

\[ g_{\mu\nu} = \left( \begin{matrix} g_{tt} & 0 & 0 & g_{t\phi} \\ 0 & g_{rr} & 0 & 0 \\ 0 & 0 & g_{\theta\theta} & 0 \\ g_{t\phi} & 0 & 0 & g_{\phi\phi} \end{matrix} \right),\]

where the only non-zero off axis elements are $g_{t\phi}$.

Required implementations:

GradusBase.inner_radius

GeodesicTracer.metric_components