Catalogue of metrics

Gradus.jl implements a library of metrics ready to use for integrations and rendering.

Gradus.AbstractMetric

Gradus.DilatonAxion

Gradus.JohannsenMetric

Gradus.JohannsenPsaltisMetric

Gradus.KerrDarkMatter

Gradus.KerrMetric

Gradus.KerrNewmanMetric

Gradus.KerrRefractive

Gradus.KerrSpacetimeFirstOrder

Gradus.MorrisThorneWormhole

Gradus.NoZMetric

Note

To implement your own custom metrics, please see Implementing a new metric. If you have a complex metric, please open an issue requesting for it to be added.

Currently available metrics

Gradus.AbstractMetric — Type

abstract type AbstractMetric{T} end

Abstract type used to dispatch different geodesic problems.

Gradus.DilatonAxion — Type

struct DilatonAxion{T} <: AbstractStaticAxisSymmetric{T}

Einstein-Maxwell-Dilaton-Axion metric.

M = 1.0: Singularity mass.

a = 0.0: Singularity spin.

β = 0.0: Dilaton coupling strength.

b = 1.0: Axion coupling strength.

Gradus.JohannsenMetric — Type

struct JohannsenMetric{T} <: AbstractStaticAxisSymmetric{T}

The Johannsen (20xx) metric.

M = 1.0: Black hole mass.

a = 0.0: Black hole spin.

α13 = 0.0: $\alpha_{13}$ deviation parameter.

α22 = 0.0: $\alpha_{22}$ deviation parameter.

α52 = 0.0: $\alpha_{52}$ deviation parameter.

ϵ3 = 0.0: $\epsilon_{3}$ deviation parameter.

Gradus.JohannsenPsaltisMetric — Type

struct JohannsenPsaltisMetric{T} <: AbstractStaticAxisSymmetric{T}

Johannsen and Psaltis 2011

M = 1.0: Black hole mass.

a = 0.0: Black hole spin.

ϵ3 = 0.0: $\epsilon_{3}$ deviation parameter.

Gradus.KerrDarkMatter — Type

struct KerrDarkMatter

https://arxiv.org/pdf/2003.06829.pdf

Gradus.KerrMetric — Type

struct KerrMetric{T} <: AbstractStaticAxisSymmetric{T}

The Kerr metric in Boyer-Lindquist coordinates, describing a black hole with mass $M$ and angular spin $a$:

\[\begin{align*} \text{d}s^2 = &- \left( 1 - \frac{2 M r}{\Sigma} \right)\text{d}t^2 - \frac{2M r a \sin^2(\theta)}{\Sigma} \text{d}t \text{d}\phi \\ &+ \frac{\Sigma}{\Delta} \text{d}r^2 + \Sigma \text{d}\theta^2 + \left(r^2 + a^2 + \frac{2 M r a^2 \sin^2(\theta)}{\Sigma} \right) \sin^2(\theta) \text{d}\phi^2, \end{align*}\]

where

\[\Sigma = r^2 + a^2 \cos^2 (\theta), \quad \text{and} \quad \Delta = r^2 - 2Mr + a^2.\]

Parameters

M = 1: black hole mass.

a = 0: black hole spin.

Gradus.KerrNewmanMetric — Type

struct KerrNewmanMetric{T} <: AbstractStaticAxisSymmetric{T}

Gradus.KerrRefractive — Type

struct KerrRefractive{T} <: AbstractStaticAxisSymmetric{T}

Kerr metric in Boyer-Lindquist coordintes with a path-length ansatz, equivalent to a refractive index n, within the coronal radius corona_radius.

M = 1.0: Black hole mass.

a = 0.0: Black hole spin.

n = 1.0: Refractive index within the corona.

corona_radius = 20.0: Radius of the corona.

Gradus.KerrSpacetimeFirstOrder — Type

A first-order implementation of KerrMetric.

M = 1.0: Black hole mass.

a = 0.0: Black hole spin.

E = 1.0: Geodesic energy (a consant of motion).

Gradus.MorrisThorneWormhole — Type

struct MorrisThorneWormhole{T} <: AbstractStaticAxisSymmetric{T}

Morris-Thorne wormhole metric.

b = 1.0: Throat size.

Gradus.NoZMetric — Type

struct NoZMetric