Accretion geometry

Gradus.jl supports the ability to implement custom accretion geometry, or even load in mesh files in any standard format using MeshIO.jl. Geometry may be standard spherically symmetric accretion discs, or any other custom type.

Note

Currently geometry is optically thick always. Radiative transfer will be added soon.

Gradus.AccretionGeometry.AbstractAccretionGeometry — Type

abstract type AbstractAccretionGeometry{T}

Supertype of all accretion geometry. Concrete sub-types must minimally implement

in_nearby_region

has_intersect

Alternativey, for more control, either intersects_geometry or collision_callback may be implemented for a given geometry type.

Geometry intersection calculations are performed by strapping discrete callbacks to the integration procedure.

Gradus.AccretionGeometry.in_nearby_region — Function

in_nearby_region(g::AbstractAccretionGeometry{T}, line_element)::Bool

Returns a boolean indicating whether the line_element is "close" to the geometry in question. Used to optimize when to call the intersection algorithm.

Contextually, "close" is a little arbitrary, and this function may always return True, however will suffer in performance if this is the case.

line_element is a tuple of two four-position vectors, indicating the last integration position, and current integration position, i.e. (u_prev, u_curr), in integrator coordinates.

Notes

This function actually depends on the step size of the integrator, but this is currently not considered in the implementation.

Gradus.AccretionGeometry.has_intersect — Function

has_intersect(g::AbstractAccretionGeometry{T}, line_element)

Returns a boolean indicating whether line_element intersects the geometry g. The intersection algorithm used depends on the geometry considered. For meshes, this uses the jsf_algorithm.

line_element is a tuple of two four-position vectors, indicating the last integration position, and current integration position, i.e. (u_prev, u_curr), in integrator coordinates.

Accretion discs

Gradus.AccretionGeometry.AbstractAccretionDisc — Type

abstract type AbstractAccretionDisc{T} <: AbstractAccretionGeometry{T}

Supertype for accretion spherically symmetric geometry, where certain optimizing assumptions may be made.

Gradus.AccretionGeometry.GeometricThinDisc — Type

struct GeometricThinDisc{T} <: AbstractAccretionDisc{T}
GeometricThinDisc(inner_radius::T, outer_radius::T, inclination::T)

inner_radius

outer_radius

inclination

Simple geometrically thin accretion disc spanning from inner_radius to outer_radius in gravitational units. Inclination of the disc is relative to spin axis, with $90^\circ$ being perpendicular to the spin axis.

Meshes