Available metrics

Gradus.jl implements a library of metrics ready to use for integrations and rendering.

Gradus.BoyerLindquistAD

Gradus.BoyerLindquistFO

Gradus.DilatonAxionAD

Gradus.JohannsenAD

Gradus.JohannsenPsaltisAD

Gradus.KerrRefractiveAD

Gradus.MorrisThorneAD

Note

To implement your own custom metrics, please see Implementing a new metric.

Catalogue of spacetimes

Gradus.BoyerLindquistAD — Type

struct BoyerLindquistAD{T} <: AbstractAutoDiffStaticAxisSymmetricParams{T}

The Kerr metric in Boyer-Lindquist coordinates, describing a black hole with mass $M$ and angular spin $a$:

\[\begin{align*} \text{d}s^2 = &- \left( 1 - \frac{2 M r}{\Sigma} \right)\text{d}t^2 - \frac{2M r a \sin^2(\theta)}{\Sigma} \text{d}t \text{d}\phi \\ &+ \frac{\Sigma}{\Delta} \text{d}r^2 + \Sigma \text{d}\theta^2 + \left(r^2 + a^2 + \frac{2 M r a^2 \sin^2(\theta)}{\Sigma} \right) \sin^2(\theta) \text{d}\phi^2, \end{align*}\]

where

\[\Sigma = r^2 + a^2 \cos^2 (\theta), \quad \text{and} \quad \Delta = r^2 - 2Mr + a^2.\]

Parameters

M

: Black hole mass. Default: 1.0

a

: Black hole spin. Default: 0.0

Gradus.BoyerLindquistFO — Type

A first-order implementation of BoyerLindquistAD.

M

: Black hole mass. Default: 1.0

a

: Black hole spin. Default: 0.0

E

: Geodesic energy (a consant of motion). Default: 1.0

Gradus.DilatonAxionAD — Type

struct DilatonAxionAD{T} <: AbstractAutoDiffStaticAxisSymmetricParams{T}

Einstein-Maxwell-Dilaton-Axion metric.

M

: Singularity mass. Default: 1.0

a

: Singularity spin. Default: 0.0

β

: Dilaton coupling strength. Default: 0.0

b

: Axion coupling strength. Default: 1.0

Gradus.JohannsenAD — Type

struct JohannsenAD{T} <: AbstractAutoDiffStaticAxisSymmetricParams{T}

The Johannsen (20xx) metric.

M

: Black hole mass. Default: 1.0

a

: Black hole spin. Default: 0.0

α13

: $\alpha_{13}$ deviation parameter. Default: 0.0

α22

: $\alpha_{22}$ deviation parameter. Default: 0.0

α52

: $\alpha_{52}$ deviation parameter. Default: 0.0

ϵ3

: $\epsilon_{3}$ deviation parameter. Default: 0.0

Gradus.JohannsenPsaltisAD — Type

struct JohannsenPsaltisAD{T} <: AbstractAutoDiffStaticAxisSymmetricParams{T}

Johannsen and Psaltis 2011

M

: Black hole mass. Default: 1.0

a

: Black hole spin. Default: 0.0

ϵ3

: $\epsilon_{3}$ deviation parameter. Default: 0.0

Gradus.KerrRefractiveAD — Type

struct KerrRefractiveAD{T} <: AbstractAutoDiffStaticAxisSymmetricParams{T}

Kerr metric in Boyer-Lindquist coordintes with a path-length ansatz, equivalent to a refractive index n, within the coronal radius corona_radius.

M

: Black hole mass. Default: 1.0

a

: Black hole spin. Default: 0.0

n

: Refractive index within the corona. Default: 1.0

corona_radius

: Radius of the corona. Default: 20.0

Gradus.MorrisThorneAD — Type

struct MorrisThorneAD{T} <: AbstractAutoDiffStaticAxisSymmetricParams{T}

Morris-Thorne wormhole metric.

b

: Throat size. Default: 1.0