Special radii

Gradus.isco — Function

isco(m::AbstractMetric)
isco(m::AbstractMetric, lower_bound, upper_bound; kwargs...)

Innermost stable circular orbit (ISCO), defined by

\[ \frac{\text{d}}{\text{d}r} \left( \frac{E}{\mu} \right) = 0.\]

Uses analytic solutions if known for that metric, else uses a root finder to calculate the radius at which the above condition is met.

Gradus.event_horizon — Function

event_horizon(m::AbstractMetric; select = last, resolution = 100, θε = 1e-7, rmax = 5.0)

Event horizon radius, often equivalent to inner_radius, however remains distinct, such that the latter may still be an arbitrary chart cutoff.

Utility function for helping plot an event horizon shape. Returns a tuple containing the r and θ vectors that solve

\[ g_{t\phi}^2 - g_{tt} g_{\phi \phi} = 0.\]

A NaN value in the r vector indicates no solution for that particular $\theta$, i.e. that the metric describes a naked singularity.

Often the equation will have multiple roots, in which case the keyword argument select may be assigned to select the desired root.