GradusBase API

Gradus.GradusBase.AbstractMetricParams — Type

abstract type AbstractMetricParams{T} end

Abstract type used to dispatch different geodesic problems.

Gradus.GradusBase.geodesic_eq — Function

geodesic_eq(m::AbstractMetricParams{T}, u, v)
geodesic_eq!(m::AbstractMetricParams{T}, u, v)

Calculate the acceleration components of the geodesic equation given a position u, a velocity v, and a metric m.

Missing docstring.

Missing docstring for geodesic_eq!. Check Documenter's build log for details.

Gradus.GradusBase.constrain — Function

constrain(m::AbstractMetricParams{T}, u, v; μ::T=0.0)

Calculate time component $v^t$ which would constrain a velocity vector v at position x as a geodesic with mass μ.

Gradus.GradusBase.inner_radius — Function

inner_radius(m::AbstractMetricParams{T})

Return the innermost valid coordinate relative to the origin, for use in geodesic tracing.

This usually represents some property of the metric, e.g. event horizon radius in Kerr/Schwarzschild metrics, or throat diameter in worm hole metrics.

Gradus.GradusBase.metric — Function

metric(m::AbstractMetricParams{T}, u)

Numerically evaluate the corresponding metric for AbstractMetricParams, given parameter values m and some point u.

Missing docstring.

Missing docstring for vector_to_local_sky. Check Documenter's build log for details.

Physical Quantities

Gradus.GradusBase.E — Function

E(m::AbstractMatrix{T}, v)
E(m::AbstractMetricParams{T}, u, v)

Compute the energy for a numerically evaluated metric, and some velocity four vector v,

\[E = - p_t = - g_{t\nu} p^\nu.\]

For null geodesics, the velocity is the momentum $v^\nu = p^\nu$. For massive geodesics, the mass $\mu$ needs to be known to compute $\mu v^\nu = p^\nu$.

Gradus.GradusBase.Lz — Function

Lz(m::AbstractMatrix{T}, v)
Lz(m::AbstractMetricParams{T}, u, v)

Compute the angular momentum for a numerically evaluated metric, and some velocity four vector v.

\[L_z = p_\phi = - g_{\phi\nu} p^\nu.\]

Geodesic Points

Missing docstring.

Missing docstring for GeodesicPoint. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getgeodesicpoint. Check Documenter's build log for details.

Missing docstring.

Missing docstring for unpack_solution. Check Documenter's build log for details.