GradusBase API documentation

Gradus.GradusBase.E — Method

E(m::AbstractMatrix{T}, v)
E(m::AbstractMetricParams{T}, u, v)

Compute the energy for a numerically evaluated metric, and some velocity four vector v,

\[E = - p_t = - g_{t\nu} p^\nu.\]

For null geodesics, the velocity is the momentum $v^\nu = p^\nu$. For massive geodesics, the mass $\mu$ needs to be known to compute $\mu v^\nu = p^\nu$.

Gradus.GradusBase.Lz — Method

Lz(m::AbstractMatrix{T}, v)
Lz(m::AbstractMetricParams{T}, u, v)

Compute the angular momentum for a numerically evaluated metric, and some velocity four vector v.

\[L_z = p_\phi = - g_{\phi\nu} p^\nu.\]

Gradus.GradusBase.constrain — Method

constrain(m::AbstractMetricParams{T}, u, v; μ::T=0.0)

Calculate time component $v^t$ which would constrain a velocity vector v at position x as a geodesic with mass μ.

Gradus.GradusBase.geodesic_eq — Method

geodesic_eq(m::AbstractMetricParams{T}, u, v)
geodesic_eq!(m::AbstractMetricParams{T}, u, v)

Calculate the acceleration components of the geodesic equation given a position u, a velocity v, and a metric m.

Gradus.GradusBase.getgeodesicpoint — Method

getgeodesicpoint(
    _::AbstractMetricParams{T},
    sol::SciMLBase.AbstractODESolution{T, N, S}
) -> FirstOrderGeodesicPoint

Used to get pack a SciMLBase.AbstractODESolution into a GeodesicPoint. Will only store information relevant to start and endpoint of the integration. To keep the full geodesic path, it is encouraged to use the SciMLBase.AbstractODESolution directly.

Gradus.GradusBase.getgeodesicpoints — Method

getgeodesicpoints(
    _::AbstractMetricParams{T},
    sol::SciMLBase.AbstractODESolution{T, N, S}
) -> Any

Unpacks each point in the solution, similar to getgeodesicpoint but returns an array of GeodesicPoint.

Gradus.GradusBase.inner_radius — Method

inner_radius(m::AbstractMetricParams{T})

Return the innermost valid coordinate relative to the origin, for use in geodesic tracing.

This usually represents some property of the metric, e.g. event horizon radius in Kerr/Schwarzschild metrics, or throat diameter in worm hole metrics.

Gradus.GradusBase.metric — Method

metric(m::AbstractMetricParams{T}, u)

Numerically evaluate the corresponding metric for AbstractMetricParams, given parameter values m and some point u.

Gradus.GradusBase.metric_components — Method

metric_components(m::AbstractMetricParams{T}, x)

Return a tuple with each non-zero metric component for the metric described by m at position x. Note that the position need not be a four-vector, and for specific implementations may only be a subset of the total manifold coordinates. See specific implementations for subtypes of AbstractMetricParams for details.

Gradus.GradusBase.metric_type — Method

metric_type(m::AbstractMetricParams{T})

Return the AbstractMetric type associated with the metric parameters m.

Gradus.GradusBase.mproject — Method

mproject(g, v, u)

Project vector v onto u with metric g. Optional first argument may be AbstractMetricParams for more optimized methods, which fallback to an einsum.

Gradus.GradusBase.AbstractGeodesicPoint — Type

abstract type AbstractGeodesicPoint

Supertype for geodesic points, used to store information about specific points along geodesic trajectories.

Note

Currently limited to storing the start and endpoint of any given trajectory. To keep the full geodesic path, it is encouraged to use the SciMLBase.AbstractODESolution directly.

Must minimally have the same fields as GeodesicPoint. Examples include Gradus.FirstOrderGeodesicPoint.

Gradus.GradusBase.AbstractMetricParams — Type

abstract type AbstractMetricParams{T} end

Abstract type used to dispatch different geodesic problems.

Gradus.GradusBase.GeodesicPoint — Type

struct GeodesicPoint <: AbstractGeodesicPoint

status: Return code of the integrator for this geodesic.

t1: Start affine time

t2: End affine time

u1: Start four position

u2: End four position

v1: Start four velocity

v2: End four velocity